Added support for iVar block parsing
Ass seen with Issue #933, the lexer fails to identify methods as after parsing the @implementation keyword as it is still in the 'statement' state and methods are only parsed in the 'root' state. The block after the class name needs to have a seperate state similar to the 'function' state so that we return to the 'root' state after exiting the block.